Frontend stuff that I already started
I got bored over reading week due to lack of access to video games and having no shows to watch, so I had already gotten started on a few of the pages that ended up being the specifications for this sprint. What a world. Good thing I didn't finish them.