Fixed Signing In and Made Updating Possible.
So, logging in wasn't working as expected and in order to actually be able to do what I needed to do I needed to be able to sign in. So, I did it myself. The user can now sign in (If they exist. I haven't touched the registry stuff but the login page already worked before this Sprint started we just didn't have the backend fully set up yet.)

Anyways, frontend page for logging in and updateable user profiles are both done because I needed to finish them.

I plan on adding a different profile view if the profile being viewed isn't yours, but as of right now, it just says you don't have permission to view it. Big day.